+++ CHANGELOG.md
@@ -13,6 +13,8 @@
 * Change text on error pages to be slightly more verbose (#302)
 * Change referrer policy on email view to same-origin (#307)
 * Use a more secure method to produce the random part of inboxes (#312)
+* Cleaned up our use of datetime to make sure we're always using timezone away
+  datetime objects
 
 ## Releases